pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "burgess-principle"
version = "1.3.0"
description = "The Burgess Principle — human scrutiny verification toolkit"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
test = ["pytest>=7.0", "PyNaCl>=1.5.0", "fastapi>=0.110.0", "uvicorn>=0.23.0", "httpx>=0.27.0", "cryptography>=3.4"]
api = ["fastapi>=0.110.0", "uvicorn>=0.23.0"]
local = ["llama-cpp-python>=0.2.0", "fastapi>=0.110.0", "uvicorn>=0.23.0"]
onchain = ["PyNaCl>=1.5.0"]

[project.scripts]
verify-scrutiny = "verify_scrutiny:main"

[tool.setuptools]
py-modules = ["verify_scrutiny", "api"]
packages = ["tracer", "iris"]